src/hu/kerdei/kata/codewars/7_kyu/Most_digits_58daa7617332e59593000006/src/MostDigits.java
import java.security.InvalidParameterException;

public class MostDigits {
    public static int findLongest(int[] numbers) {

        if (numbers == null || numbers.length == 0) {
            throw new InvalidParameterException();
        }

        int longestDigit = numbers[0];

        for (int i = 1; i < numbers.length; i++) {
            if (String.valueOf(Math.abs(longestDigit)).length() < String.valueOf(Math.abs(numbers[i])).length()) {
                longestDigit = numbers[i];
            }
        }

        return longestDigit;

    }
}